Speed up tests, without causing Spring errors
It appears that Spring currently has some problems interacting with
Zeitwerk, the new autoloader that was added in Rails 6.

The recommended workaround has caused `rails new` to set
`config.cache_classes = false` by default, which causes a test suite to
run very slowly, especially when Capybara is being used.

An alternative workaround was mentioned in the Spring repo
(https://github.com/rails/spring/issues/598), which allows
`cache_classes = true` to be set. The compromise is that it uses the
'classic' autoloader instead of Zeitwerk.