Test database can be created with the default postgres user
WIthout this change the user is required to create a new role in their local postgres database with something similar to: `createuser --super test`.

We can avoid this hurdle by using the default postgres role called `postgres` which is what's used in the development configuration. The separate database names for `rails-template-development` and `rails-template-test` should continue to provide enough separation.