Remove Selenium deprecation warning
Before this change, running the test suite would produce the following deprecation warning:

```
2020-08-17 17:42:57 WARN Selenium [DEPRECATION] Selenium::WebDriver::Chrome#driver_
path= is deprecated. Use Selenium::WebDriver::Chrome::Service#driver_path= instead.
```

After looking it up it seems to be coming from the chromedriver-helper gem. In following this advice I've removed it and it continues to function:

https://github.com/SeleniumHQ/selenium/issues/7125#issuecomment-486917791

As an additional check I looked at BEIS ODA to see how an app with many more tests is configured. It too has Selenium and has removed this chromedriver-helper gem. This is interesting because that repo was creating using this template.