Install Simplecov to generate code coverage
This gem will produce a code coverage report in the .lcov format. Placing it in the default file location of ./coverage/lcov/app.lcov.

This gem is being used instead of the Coveralls gem as it does not produce results in the .lcov format required by the GitHub Action we can reuse by Coveralls.

We can instruct the formatter gem to produce one file rather than many. Something you can't do with Coveralls.