CI will check the code coverage
As the last step of the test action the GitHub Action called Coveralls will run and report the new coverage to coveralls.io. This will then report back whether this is acceptable or not in a webhook back to the pull request.

These settings can be configured with coverage thresholds in Coveralls itself once authenticated with GitHub:
https://coveralls.io/github/dxw/rails-template/settings