Run CI workflow on merges to main or develop
Previously, CI would only run on pull requests, and not on the result of merging those pull requests. This would mean that downstream automated processes relying on a passing CI result (such as Heroku deployments) would never run.

This change will run CI workflows on all pull requests, and all pushes (including the result of merges) to `develop` and `main`.