Add an automatic rebase action
Sometimes, we set up projects to require the branches to be up to date before
merging. With lots of devs on one project at a time and PRs flying everywhere, this
can get annoying having to pull down and rebase code.

This adds an action to allow users to add a comment with the `/rebase` command,
the code then gets pulled down, rebased and pushed back up.

The only gotcha is that, in order for CI to be kicked off again by the push, we need
the repo to have a personal access token set. I've added a check for this, which
adds a note to the PR if the repo has a missing access token, with instructions on
how to fix this.