Tidy up the gitignore file
Rails inserts a lot of cruft that's either not relevant for us or
belongs in an individual's global gitignore file. Long, complicated
gitignore files are harder to read and understand, and make it hard to
make decisions about when it's wrong or why a thing is ignored.

Instead, we should keep it terse and focused on the things we _know_ we
need to ignore, and add to it when we need to.