Create a .dockerignore file from .gitignore
Most of the rules will be the same in both files, but there are some
syntax differences and additional rules or exceptions to make for a
Docker image, so we can't use a symlink. Instead we add some
documentation guidance to remind people to update both places and about
the differences between them.