Prevent accidental disclosure of sensitive files during development
Using `COPY . .` is problematic as it could (and has) quietly included sensitive files such as `.env` which was then pushed to a Docker registry.

Instead of doing this we can change to a whitelist approach. Including only the  folders we expect for a Rails project. These should remain static and perhaps only change during major Rails releases.

A `.dockerignore` file does already exist as a mechanism to achieve the same result however it requires a human to rememeber to configure it properly -  steps have also been made to improve this file by default [1]. This Dockerfile change is therefore aimed at provided more defence in depth to help prevent accidental disclosure of sensitive files.

We are using this in production already at least on DfE Schools Buying [2] and DfE Claim [3].

We have to copy in the dotfiles for our linters so they're available. Next step is to use docker layers to isolate these dependencies from the final build image that's run in prod.

[1] https://github.com/dxw/rails-template/pull/198
[2] https://github.com/DFE-Digital/buy-for-your-school/blob/develop/Dockerfile#L33
[3] https://github.com/DFE-Digital/claim-additional-payments-for-teaching/blob/master/Dockerfile#L101