Starting the app with docker will be automatically bundle if needed
When the gem file changes and we start the server through docker, it will correctly complain that the image needs rebuilding. It's doing this because it finds a change in the gemfile.lock.

The problem is that this is jarring, as the developer needs to interupt their chain of thought and remember what to do: first do `docker-compose build` before running `script/dserver` again. This can take some time so it's easy to forget where you were.

This change adds a little sprinkle of convenience. When the server starts in development it will check if it needs to bundle one final time before starting the server. This still takes a little time to bundle but the server starts cleanly.

It is not a replacement for `docker-compose build` but it will allow you to do this separately at a later time. I often start open a new terminal when I see this happen and run `docker-compose build` and forget about it.