Docker compose is used by CI to test the application
This codifies the knowledge of how to start a set of Docker containers into a reusable docker compose format.

Not only can CI do this but we can also use it in local development work whilst maintaining a single definition. This means we will be able to work with the same environment configuration as CI which is helpful to catching bugs early and replicating issues found later on.