CI docker image caching is less fragile
When Docker Compose builds images it will use the name of the current directory (currently Rails Template) and suffix the name of the container (test).

This will create a docker image: rails-template_test

That image name is great for the template itself but when it's used as a template a new directory name will be given, for example MyGreatApp.

CI uses `docker save` with rails-template_test for caching. If we did nothing then Docker Compose would build against an image of a different name and not be able to take advantage of the inherent advantages of building from cached docker layers.

This change removes the needs to add another TODO for the team to change this when using the template by using the Docker Compose project name argument[1]

[1] https://docs.docker.com/compose/reference/#use--p-to-specify-a-project-name