Enable Rails Content Security Policy
We don't want scripts or images being loaded from external sites.

A defence in depth step. If user input is manipulated without being sanitised it could be allowed to make requests for external assets. For example, if the user  could inject a `<img src="evil.com">` into their or another users user-agent then they could load or execute harmful or information revealing scripts, improper images etc.

Turning this policy on tells user-agents not to try and load content from anywhere that's not the current domain(self) and HTTPS.

This should be our default standpoint as it is picked up in pen tests. We should have a good reason for turning this setting off, which applications are still free to do.

Additional trusted hosts can be added to this configuration as they become necessary.

https://guides.rubyonrails.org/security.html#content-security-policy