Install Webpacker
The changes in this commit are the result of adding `webpacker` to
Gemfile and running `rails webpacker:install`. Most of the code here is
generated by that command, it's not hand-written, and I have not changed
any of the default configuration.

Webpacker [1] is a gem that integrates Webpack [2] into Rails, allowing
JavaScript and other front-end assets to be compiled using Webpack
instead of the Rails asset pipeline. In recent Rails releases it's the
default way of handling JavaScript.

The reason to use this toolchain instead of the Rails asset pipeline
that some of the tools in that pipeline don't fully support modern
JavaScript syntax. Moving the Webpack and Babel [3] means we can write
any modern JS syntax and have it compiled down to code that's compatible
with browsers we need to support under GOV.UK guidelines, for example
IE11 [4].

[1]: https://github.com/rails/webpacker
[2]: https://webpack.js.org/
[3]: https://babeljs.io/
[4]: https://www.gov.uk/service-manual/technology/designing-for-different-browsers-and-devices