Install CoffeeScript support for Webpacker
Since we have some CoffeeScript code in this app, we need Webpack to run
the CoffeeScript compiler over some source files when they're imported
from other JS modules. We do this by running:

    rails webpacker:install:coffee

The changes here are the result of that command. Note that I had to
downgrade `coffee-loader` to 1.0, because using 2.0 causes modules to
fail to load.