Update ESLint config to handle Webpacker
Files generated by the Rails Webpacker installer, or by the Webpack
build process, should not be linted -- only application code we write
should be checked. We add entries to `.eslintignore` so that config
files and compiler output are not linted.

We also need to add some rules to handle Node-style and ECMAScript
standard modules. The former uses the `require()` function to load
modules and the latter uses the `import` and `export` keywords. The
settings added to `.eslintrc` allow ESLint to handle these expressions
without errors.