Expose the git SHA and the build time of the last build in `/healthcheck`
During a recent hackday, an idea was put forward to expose the git sha and
build time of a deployed application's current deployment. As a team, we worked
to expose these details in the `/healthcheck` endpoint of an existing
application. This work is replicated here in rails-template to make it a
part of our boilerplate application.

As rails-template does not include a build script, this piece is missing from
this commit. To populate these values, pass them as arguments when building
the docker image for deployment. For example:

```
docker build --target web \
  --build-arg current_sha="$GITHUB_SHA" \
  --build-arg time_of_build="$(date -u +'%Y-%m-%dT%H:%M:%SZ')" \
  -t "<where the image is to be deployed>:$DOCKER_TAG" \
  .
```

Assuming you are using Github actions to deploy, the sha is grabbable from
${GITHUB_SHA} in github actions.

To view a demo application consuming this data, see:
https://github.com/dxw/detect-a-deployment-hackday

Co-authored-by: lawrence-forooghian <lawrence@dxw.com>
Co-authored-by: pezholio <stuart@dxw.com>
Co-authored-by: leedxw <lee@dxw.com>