Upgrade Node to active version
We noticed that version 12 was coming close to EOL. It also wasn't on the latest security patch.

This change bumps it to latest active version 16.9.1 which will soon before the active version.

For the tests to pass, we need to ignore the coverage directory from the js tests as there's a lot of files in there we don't want to try and correct. Without this change the tests fail.

This removes a warning from new docker builds where it would hang for 20 seconds to let you read an EOL warning.