Fix running of specific tests through script/test
`--pattern` is not the conventional way we tell rspec what files to run. As per the documentation [1] we can simply pass files or paths for multiple files in as a parameter.

Using the `--pattern` control presents the user with a load file error.

```
$ bundle exec rspec --pattern spec
An error occurred while loading ./spec.
Failure/Error: return load_without_bootsnap(resolved, wrap)

LoadError:
  cannot load such file -- /Users/x/sites/dxw/rails-template/spec
```

[1] https://rspec.info/documentation/4.0/rspec-rails/#Running_specs