Prevent setting this app from changing global bundle
When we bootstrap we get this message:

```
==> Bootstrapping...
==> Installing Ruby dependencies...
Your application has set path to "vendor/bundle". This will override the global value you are currently setting
```

This points at the change to your systems `bundle config`. If you check it it will have set the global path to vendor/bundle.

This will cause problems as you switch between different repositories and it'll stop using system gems without you knowing. This can cause instability as we saw on CITA with Webpacker failing spectacularly.