By default the Dockerfile only installs production gems
`bundle config set with $BUNDLE_GEM_GROUPS` wasn't doing what we thought in the production context.

It'd work out to be `bundle config set with production` which would bundle everything, rather than only production groups.
I'm not sure where this assumption came from. 'default' also doesn't work.

As the first step of a bit of a refactoring of the Dockerfile, we cement the assurance that by default the dockerfile _only_ bundles production gems.
We remove the way the build_arg `RAILS_ENV` was used to accomodate being built for production, development and test. This hopefully makes it much clearer.

(I've also set NODE_ENV to follow the same path as RAILS_ENV -- Dragon)