Dockerfile installs development and test gems in unique build stage
Instead of passing `RAILS_ENV` as a build argument to the dockerfile to figure out what should be bundled.
We switch to using `--target` to target a specific docker build stage that hardcodes the `RAILS_ENV` so it can't have an unexpected value.

We now have a stage for production, development and test. Each explicitly installs gems for the right group.