Application starts in production mode
When starting the application with RAILS_ENV=production mode and we run `rake` we get an error that explains we don't have standard installed.

This is correct as out gemfile specifies that we only install standard in the dev and test environments.

This change only tries to load the standard rake task when in the dev or test environment.

Other take tasks such as `rails db:prepare` can still run in any environment as normal.