Fix or ignore stylelint errors
This fixes errors reported by stylelint after its introduction in
the last few commits. One issue it raised was the use of `@extend`
with classes rather than placeholder selectors. Since we're
extending Bootstrap classes here, we don't have the option to use
placeholder selectors, so I've disabled this rule in the offending
stylesheet

@see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-extend-no-missing-placeholder/README.md