Update spring gem for Rails 7
We needed to go from v2 to at least v3.0.0 as per:

https://edgeguides.rubyonrails.org/upgrading_ruby_on_rails.html#upgrading-from-rails-6-1-to-rails-7-0-spring

Removing `spring-watcher-listen` was the easy way to do
this. If that's an important part of people's workflow then
they can figure out how to make it work. (I note from
https://github.com/rails/spring-watcher-listen that there
may be issues running spring-watcher-listen with ARM Macs.)