(Fix) Update Dockerfile for Ruby 3.2.2 upgrade
We had some CI failures on the Ruby 3.2.2 Renovate PR. The CI failed like so:

```
==> Linting CSS...
yarn run v1.22.19
$ npx stylelint "app/**/*css"
/bin/sh: 1: npx: not found
error Command failed with exit code 127.
```

After a bit of Googling the problem seemed to be a node version which did not
support the npx command (a bug). Updating the node version in the Dockerfile
seems to fix this.

At the same time, update the Rubygems version and Bundler versions, as
maintenance.