Revert "Compile assets in Docker-less setup and update scripts"
This reverts commit b33baf545f704bdf8f82615d1f062b1a05dc9ee7.

These scripts try to run `yarn build` and `yarn build:css`. It looks
like I was trying to port a fix for a missing assets issue over from the
Apply for Landing app to the Rails template, but didn't spot that these
scripts aren't even defined in the Rails template's package.json. Oh
dear

I didn't encounter any asset-related issues when building this app
afresh recently, but if it comes up in future, it might be worth looking
at those scripts in the Apply for Landing app